fix(api): harden the republisher walk, record cache, and worker-loop seam
Address the code-review findings on the in-process republisher (#634):

- record_cache.sequence is numeric(20,0), not signed bigint, so a uint64
  sequence >= 2^63 (garbage/hostile network answer) stores instead of
  faulting the upsert and aborting the whole sweep. Migration and entity
  moved together; Migration Drift stays green.
- A re-PUT record whose sequence is unreadable now caches at the floor 0
  instead of skipping, so a successful re-PUT is never orphaned by a
  0-row markRepublished (which would 404 recovery and hide the name from
  staleness alerts). CipherBox records carry sequence >= 1, so the floor
  only bites non-CipherBox/garbage bytes and the monotonic guard keeps a
  later real record winning.
- The walk short-circuits when no routing endpoint is configured
  (RecordTransport.configured), so a BYO-only deploy no longer emits an
  every-name resolve-failure alert storm each cadence.
- TimerWorkerScheduler bounds every run with a per-run timeout
  (PeriodicTask.runTimeoutMs): a wedged sweep is abandoned, clears the
  in-flight flag, and can neither drop all later ticks nor hang shutdown.
- The walk runs names through a bounded worker pool
  (REPUBLISHER_CONCURRENCY) so one slow name can't serialize the sweep
  past the cadence.
- The REPUBLISHER_ENABLED opt-out parses case-insensitively
  (false/0/no/off), not only the exact string "false".

+++ apps/api/src/migrations/1784600557946-AddRecordCache.ts
@@ -5,7 +5,7 @@ export class AddRecordCache1784600557946 implements MigrationInterface {
 
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.query(
-      `CREATE TABLE "record_cache" ("ipns_name" character varying(128) NOT NULL, "record" bytea NOT NULL, "sequence" bigint NOT NULL, "last_republished_at" TIMESTAMP WITH TIME ZONE, "created_at" TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT now(), CONSTRAINT "PK_49f183426ca65b21642fb75178c" PRIMARY KEY ("ipns_name"))`
+      `CREATE TABLE "record_cache" ("ipns_name" character varying(128) NOT NULL, "record" bytea NOT NULL, "sequence" numeric(20,0) NOT NULL, "last_republished_at" TIMESTAMP WITH TIME ZONE, "created_at" TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT now(), CONSTRAINT "PK_49f183426ca65b21642fb75178c" PRIMARY KEY ("ipns_name"))`
     );
   }
